requirements.txt
azure-core==1.30.1
azure-functions==1.19.0
azure-storage-blob==12.20.0
pandas==2.2.2
pyodbc
pillow==10.3.0
sqlalchemy
tensorflow-cpu